requirements_dev.txt
ansible-compat==24.10.0
ansible-core==2.14.1
ansible-lint==6.22.2
attrs==25.1.0
black==25.1.0
bracex==2.5.post1
cffi==1.17.1
cfgv==3.4.0
click==8.1.8
cryptography==44.0.1
distlib==0.3.9
filelock==3.17.0
grimp==3.9
identify==2.6.8
import-linter==2.3
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.6.0
pre_commit==4.1.0
pycparser==2.22
Pygments==2.19.1
pytest==8.4.1
PyYAML==6.0.2
referencing==0.36.2
resolvelib==0.8.1
rich==13.9.4
rpds-py==0.23.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.12.7
subprocess-tee==0.4.2
tomli==2.2.1
typing_extensions==4.12.2
virtualenv==20.29.2
wcmatch==10.0
yamale==6.0.0
yamllint==1.35.1
